rename entities to SubstrateEventEntity and SubstrateExtrinsicEntity
